+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Rely on `gha-utils labels` to dump label configuration files for inspection and bootstrapping.
+
 ## [5.2.0 (2026-01-23)](https://github.com/kdeldycke/workflows/compare/v5.1.0...v5.2.0)
 
 - Add `gha-utils labels` command to dump bundled label configuration files.